Update drifting fad array

+++ useful_scripts/prep_drifting_fadArray.R
@@ -22,10 +22,12 @@ interpolation_method = "linear"
 
 timestep = 100 #seconds
 path_duration = 120 #days
-n_steps_per_file = 200
+n_steps_per_file = 10000
 
 # !!! must be equal to N * timesteps with N integer !!!
 interpolation_time = 12 #hours
+
+RESET = T
 
 tuna_release_date = as.POSIXct("2020-01-01 12:00:00")
 
@@ -33,6 +35,9 @@ tuna_release_date = as.POSIXct("2020-01-01 12:00:00")
 tuna_position_date <- seq(from = tuna_release_date,
                           to = tuna_release_date + as.difftime(path_duration, units = "days"),
                           by = timestep)
+tuna_position_date %>% as.data.frame() %>% mutate(timestep = seq(from = 0,
+                                                                 by = 1,
+                                                                 length.out = length(tuna_position_date))) -> total_timesteps
 
 ### On reconstruit les trajectoires en prenant en plus les donnees 12h avant et apres les dates d'interets
 ### pour etre surs d'avoir tous les points au moment du depart
@@ -48,27 +53,60 @@ steps_for_files <- seq(from = 0, to = length(tuna_position_date)-1, n_steps_per_
 dir.create(file.path(OUTPUT_PATH, "drifting_fadArray"), showWarnings = F) #create the main directory
 
 existing_files <- list.files(file.path(OUTPUT_PATH, "drifting_fadArray")) #if the directory is not empty, delete files
-unlink(file.path(OUTPUT_PATH, "drifting_fadArray", existing_files))
+if (RESET == T){
+  unlink(file.path(OUTPUT_PATH, "drifting_fadArray", existing_files))
+}
 
-header = c("longitude","latitude","position_date","timestep", "FAD_id","\n")
+
+# ~~~~ Sub-function to save files
+var_names <- c("x","y","id")
+col_names <- c("longitude", "latitude", "FAD_id")
+
+save.csv.traj <- function(var_names, col_names, k, steps_for_files){
+  
+  interp_data %>%
+    dplyr::filter(timestep >= steps_for_files[k] & timestep < steps_for_files[k+1]) -> data.k
+  
+  for (i in 1:length(var_names)){
+    write.table(round(t(data.k[,col_names[i]]), digits = 5),
+                file = file.path(OUTPUT_PATH, "drifting_fadArray",
+                                 paste0("FADs.",var_names[i],"_", format(steps_for_files[k], scientific = F),
+                                        "to",
+                                        format(steps_for_files[k+1]-1, scientific = F),
+                                        ".csv")),
+                sep = ",", append = T, row.names = F, col.names = F
+    )
+  }
+  
+  
+}
+
+
+# header = c("longitude","latitude","position_date","timestep", "FAD_id","\n")
+
 
 # create the output files
-for (k in 1:(length(steps_for_files)-1)){
-  cat(paste(header, collapse = ","),
-      file = file.path(OUTPUT_PATH, "drifting_fadArray",
-                        paste0(format(steps_for_files[k], scientific = F),
-                               "to",
-                               format(steps_for_files[k+1]-1, scientific = F),
-                               ".csv"))
-      )
-}
-cat(paste(header, collapse = ","),
-    file = file.path(OUTPUT_PATH, "drifting_fadArray",
-                     paste0(format(steps_for_files[length(steps_for_files)], scientific = F),
-                            "to",
-                            format(length(tuna_position_date), scientific = F),
-                            ".csv"))
-)
+# for (j in 1:length(var_names)){
+# for (k in 1:(length(steps_for_files)-1)){
+#   
+#       file.create(file.path(OUTPUT_PATH, "drifting_fadArray",
+#                         paste0(var_names[j],format(steps_for_files[k], scientific = F),
+#                                "to",
+#                                format(steps_for_files[k+1]-1, scientific = F),
+#                                ".csv")))
+#       
+# }
+# 
+#     file.create(file.path(OUTPUT_PATH, "drifting_fadArray",
+#                      paste0(var_names[j],format(steps_for_files[length(steps_for_files)]-1, scientific = F),
+#                             "to",
+#                             format(length(tuna_position_date), scientific = F),
+#                             ".csv"))
+# )
+# }
+# file.create(file.path(OUTPUT_PATH, "drifting_fadArray", "FADs.x.csv"),
+#             file.path(OUTPUT_PATH, "drifting_fadArray", "FADs.y.csv"),
+#             file.path(OUTPUT_PATH, "drifting_fadArray", "FADs.id.csv"))
 
 
 #Initialize progress bar
@@ -83,12 +121,14 @@ names(buoysSummary) <- c("buoy_id", "FAD_id", "is_notEmpty", "is_inAreaOfInteres
 
 for (i in 1:length(buoysList)){
   
+  # tick progress bar
   pb$tick(tokens = list(id_buoy = buoysList[i],
                         current = i,
                         n_buoys = length(buoysList)))
   
   if( file.exists(file.path(DATA_PATH,buoysList[i],"buoys_data_location_filtered.RData")) ){
     
+    # load RData
     load(file.path(DATA_PATH,buoysList[i],"buoys_data_location_filtered.RData"))
     
     if(dim(buoysDataLocation)[1] <= 1){
@@ -129,7 +169,7 @@ for (i in 1:length(buoysList)){
             end_fad_path <- max(interpolation_date)
           }
           
-          
+          # timesteps of interpolation
           interp_t <- seq(from = start_fad_path, to = end_fad_path, by = timestep)
           
           #interpolate x and y positions
@@ -147,39 +187,55 @@ for (i in 1:length(buoysList)){
                                   by = 1, length.out = length(position_date)),
                    FAD_id = i) %>%
             # change timestep so that it starts from 0 (for the model in Python)
-            mutate(timestep = timestep - 1) -> interp_data
+            mutate(timestep = timestep - 1) %>%
+            right_join(total_timesteps, by = "timestep") %>%
+            select(longitude, latitude, position_date, timestep, FAD_id) -> interp_data
           
           # Fill in summary
           buoysSummary[i,] <- c(buoysList[i], i, T, T, T)
           
           # Save positions
           for (k in 1:(length(steps_for_files)-1)){
-            interp_data %>%
-              dplyr::filter(timestep >= steps_for_files[k] & timestep < steps_for_files[k+1]) -> data.k
-            if (dim(data.k)[1] > 0){
-              write.table(data.k,
-                          file = file.path(OUTPUT_PATH, "drifting_fadArray",
-                                           paste0(format(steps_for_files[k], scientific = F),
-                                                  "to",
-                                                  format(steps_for_files[k+1]-1, scientific = F),
-                                                  ".csv")),
-                          sep = ",", append = T, row.names = F, col.names = F
-              )
-            }
+              
+              save.csv.traj(var_names, col_names, k, steps_for_files)
+          
           }
+          
+          
           interp_data %>%
             dplyr::filter(timestep >= steps_for_files[length(steps_for_files)] & timestep <= length(tuna_position_date)-1 ) -> data.k
-          if (dim(data.k)[1] > 0){
-            write.table(data.k,
+          
+          for (i in 1:length(var_names)){
+            write.table(round(t(data.k[,col_names[i]]), digits = 5),
                         file = file.path(OUTPUT_PATH, "drifting_fadArray",
-                                         paste0(format(steps_for_files[length(steps_for_files)], scientific = F),
+                                         paste0("FADs.",var_names[i],"_",
+                                                format(steps_for_files[length(steps_for_files)], scientific = F),
                                                 "to",
                                                 format(length(tuna_position_date)-1, scientific = F),
                                                 ".csv")),
                         sep = ",", append = T, row.names = F, col.names = F
             )
-            
           }
+          
+          
+            
+            
+
+          
+          
+          # Save positions
+         # write.table(round(t(interp_data$longitude), digits = 5),
+         #             file = file.path(OUTPUT_PATH, "drifting_fadArray", "FADs.x.csv"),
+         #             sep = ",", append = T, row.names = F, col.names = F)
+         # write.table(round(t(interp_data$latitude), digits = 5),
+         #             file = file.path(OUTPUT_PATH, "drifting_fadArray", "FADs.y.csv"),
+         #             sep = ",", append = T, row.names = F, col.names = F)
+         # write.table(round(t(interp_data$FAD_id), digits = 5),
+         #             file = file.path(OUTPUT_PATH, "drifting_fadArray", "FADs.id.csv"),
+         #             sep = ",", append = T, row.names = F, col.names = F)
+         #  
+          
+          
           
           # ggplot()+geom_path(data = interp_data, aes(x = longitude, y = latitude, color = timestep))
